fix that allows up to 255 cclients on the server

+++ src/netcom/socket_templates.h
@@ -56,7 +56,7 @@ typedef struct io_handler {
 
 	struct message_byte_array *buffer;
 	uint8_t *read_buffer;
-	uint8_t recv_length;
+	uint16_t recv_length;
 	int status;
 
 	int sfd_read_write;
